ENH: Drop Python 3.7 and add 3.10 to testing matrix
Drop Python 3.7 and add 3.10 to testing matrix: the end-of-life of 3.7
is expected in 6 months (27 Jun 2023):
https://devguide.python.org/versions/#supported-versions

Enclose Python version numbers in quotes, as otherwise 3.10 is taken as
3.1 as otherwise YAML parses them as numbers and trims 3.10 to 3.1:
https://github.com/carpentries-incubator/SDC-BIDS-dMRI/actions/runs/3773615353

Documentation:
https://github.com/actions/setup-python/blob/main/docs/advanced-usage.md#using-the-python-version-input

Upgrade to DIPY 1.5.0 as previous versions are not compatible with
Python 3.10.

Specify a Python version for Binder through the `runtime.txt` file. The
current default Python version for Binder is 3.7, and hence the
notebooks would fail unless a newer version is specified as we are
dropping support for Python 3.7.

Documentation:
https://mybinder.readthedocs.io/en/latest/howto/languages.html#python