DOC: Updated preprocessing and introduction modules
Added in and expanded upon the current information in the introduction
and preprocessing lessons. This includes changes to improve clarity and
the logical flow of the lesson. Also changed the first figure of the
introduction lesson to something that is a bit more clear (i.e. easier
to understand the signal change via gradient direction change).